HINT_COUNTER || Updated Hint function.
When typing "hint" in the answer bar, program would continue to respond with hints even when the MAX amount of hints had been used (MAX: 3). I have updated this bug and fixed it to respond with "you have run out of hints" for when the player enters "hint" when all hints have already been used.